Move createDefaultUserRelations to core file
