+++ src/frontend/packages/cloud-foundry/src/actions/organization.actions.ts
@@ -1,3 +1,5 @@
+import { HttpHeaders, HttpRequest } from '@angular/common/http';
+
 import { IUpdateOrganization } from '../../../core/src/core/cf-api.types';
 import { getActions } from '../../../store/src/actions/action.helper';
 import { PaginatedAction } from '../../../store/src/types/pagination.types';
@@ -17,8 +19,7 @@ import {
   EntityInlineParentAction,
 } from '../entity-relations/entity-relations.types';
 import { CFStartAction } from './cf-action.types';
-import { createDefaultUserRelations } from './user.actions.helpers';
-import { HttpRequest, HttpHeaders } from '@angular/common/http';
+import { createDefaultUserRelations } from './users.actions';
 
 export const GET_ORGANIZATION = '[Organization] Get one';
 export const GET_ORGANIZATION_SUCCESS = '[Organization] Get one success';
