--- src/frontend/packages/cloud-foundry/src/actions/user.actions.helpers.ts
@@ -1,15 +1,0 @@
-import { cfUserEntityType } from '../cf-entity-types';
-import { CfUserRoleParams } from '../store/types/user.types';
-import { createEntityRelationKey } from '../entity-relations/entity-relations.types';
-
-export function createDefaultUserRelations() {
-  return [
-    createEntityRelationKey(cfUserEntityType, CfUserRoleParams.ORGANIZATIONS),
-    createEntityRelationKey(cfUserEntityType, CfUserRoleParams.AUDITED_ORGS),
-    createEntityRelationKey(cfUserEntityType, CfUserRoleParams.MANAGED_ORGS),
-    createEntityRelationKey(cfUserEntityType, CfUserRoleParams.BILLING_MANAGER_ORGS),
-    createEntityRelationKey(cfUserEntityType, CfUserRoleParams.SPACES),
-    createEntityRelationKey(cfUserEntityType, CfUserRoleParams.MANAGED_SPACES),
-    createEntityRelationKey(cfUserEntityType, CfUserRoleParams.AUDITED_SPACES)
-  ];
-}
